example_project/requirements.txt
beautifulsoup4==4.10.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2 
Django==4.0.2
djangorestframework==3.13.1
requests==2.27.1
bleach==4.1.0
django-extensions==3.1.5
django-model-utils==4.2.0
django-filter==21.1
ordered-set==4.0.2
django-environ==0.8.1
git+https://${GITHUB_TOKEN}@github.com/digirati-co-uk/drf-search-service.git@develop
